// src/types/editor.ts
export type ITrackType = "main" | "text" | "image" | "video" | "audio" | "helper" | "caption" | "template" | "composition" | "illustration" | "shape" | "rect" | "progressBar" | "progressSquare" | "progressFrame" | "radialAudioBars" | "linealAudioBars" | "waveAudioBars" | "hillAudioBars";
export type ITransitionType = "none" | "fade" | "slide" | "wipe" | "flip" | "clockWipe" | "star" | "circle" | "rectangle" | "slidingDoors";
export type IKindHistory = "add" | "remove" | "update" | "replace" | "update:details" | "layer:selection" | "undo" | "design:resize" | "design:load" | "redo" | "add:transition" | "edit:track";
export interface ISize {
    width: number;
    height: number;
    type?: string;
    name?: string;
}
export interface IDisplay {
    from: number;
    to: number;
}
export interface ITrim {
    from: number;
    to: number;
}
export interface IBoxShadow {
    color: string;
    x: number;
    y: number;
    blur: number;
}
export interface IMetadata {
    resourceId: string;
    order: number;
}
// ...
export interface ITransition {
    id: string;
    trackId: string;
    fromId: string;
    toId: string;
    type: string;
    name?: string;
    duration: number;
    preview?: string;
    direction?: any;
    kind: string;
}
export interface ITrack {
    id: string;
    type: ITrackType;
    items: string[];
    muted?: boolean;
    metadata?: Partial<IMetadata>;
    accepts?: string[];
    index?: number;
    magnetic?: boolean;
    static?: boolean;
}
export interface ITimelineScaleState {
    unit: number;
    zoom: number;
    segments: number;
    index: number;
}
export interface ITimelineScrollState {
    /**
     * Timeline scroll state by X-axis.
     */
    left: number;
    /**
     * Timeline scroll state by Y-axis.
     */
    top: number;
}
export interface CanvasSpacing {
    left: number;
    right: number;
}
export interface ItemStructure {
    id: string;
    items: string[];
    transitions: string[];
    tracks: ITrack[];
}
// ...
export type ITrackItem = IAudio | IImage | IText | IVideo | ICaption | ITemplate | IComposition | IIllustration | IShape | IRect | IProgressBar | IProgressSquare | IProgressFrame | IRadialAudioBars | ILinealAudioBars | IWaveAudioBars | IHillAudioBars;
// ...
export interface IText extends ITrackItemBase {
    type: "text";
    details: ITextDetails;
}
export interface ICaption extends ITrackItemBase {
    type: "caption";
    details: ICaptionDetails;
}
export interface IImage extends ITrackItemBase {
    type: "image";
    details: IImageDetails;
}
export interface IVideo extends ITrackItemBase {
    type: "video";
    details: IVideoDetails;
}
export interface IAudio extends ITrackItemBase {
    type: "audio";
    details: IAudioDetails;
}
export interface IIllustration extends ITrackItemBase {
    type: "illustration";
    details: IIllustrationDetails;
}
export interface IShape extends ITrackItemBase {
    type: "shape";
    details: IShapeDetails;
}
export interface IRect extends ITrackItemBase {
    type: "rect";
    details: IRectDetails;
}
export interface IProgressBar extends ITrackItemBase {
    type: "progressBar";
    details: IProgressBarDetails;
}
export interface IProgressSquare extends ITrackItemBase {
    type: "progressSquare";
    details: IProgressSquareDetails;
}
export interface IProgressFrame extends ITrackItemBase {
    type: "progressFrame";
    details: IProgressFrameDetails;
}
export interface IRadialAudioBars extends ITrackItemBase {
    type: "radialAudioBars";
    details: IRadialAudioBarsDetails;
}
export interface ILinealAudioBars extends ITrackItemBase {
    type: "linealAudioBars";
    details: ILinealAudioBarsDetails;
}
export interface IWaveAudioBars extends ITrackItemBase {
    type: "waveAudioBars";
    details: IWaveAudioBarsDetails;
}
export interface IHillAudioBars extends ITrackItemBase {
    type: "hillAudioBars";
    details: IHillAudioBarsDetails;
}
export interface IComposition extends ITrackItemBase {
    type: "composition";
    trackItemIds: string[];
    trackItemsMap: Record<string, ITrackItem>;
    tracks: ITrack[];
    size: ISize;
}
export interface ITemplate extends ITrackItemBase {
    type: "template";
    trackItemIds: string[];
    trackItemsMap: Record<string, ITrackItem>;
    transitionsMap: Record<string, ITransition>;
    transitionIds: string[];
    size: ISize;
    tracks: ITrack[];
    structure: ItemStructure[];
}
// ...
export interface IDesign {
    id: string | number;
    size: ISize;
    duration?: number;
    fps: number;
    tracks: ITrack[];
    trackItemIds: string[];
    trackItemsMap: Record<string, ITrackItem>;
    transitionIds: string[];
    transitionsMap: Record<string, ITransition>;
    structure?: ItemStructure[];
    background?: {
        type: "color" | "image";
        value: string;
    };
}